requirements.txt
pillow>=8.2.0
requests>=2.23.0
praw>=7.1.0
instabot>=0.117.0
validit[yaml]==1.3.1
simplejsondb==0.4.0
termcolor
dataclasses; python_version < "3.7"
